[AIP-21] Move contrib.hooks.spark_sql_hook providers.apache.spark.hooks.spark_sql
