+++ airflow/contrib/operators/spark_sql_operator.py
@@ -17,8 +17,8 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-from airflow.contrib.hooks.spark_sql_hook import SparkSqlHook
 from airflow.models import BaseOperator
+from airflow.providers.apache.spark.hooks.spark_sql import SparkSqlHook
 from airflow.utils.decorators import apply_defaults
