+++ tests/test_core_to_contrib.py
@@ -164,6 +164,10 @@ HOOK = [
         'airflow.providers.apache.spark.hooks.spark_jdbc.SparkJDBCHook',
         'airflow.contrib.hooks.spark_jdbc_hook.SparkJDBCHook',
     ),
+    (
+        'airflow.providers.apache.spark.hooks.spark_sql.SparkSqlHook',
+        'airflow.contrib.hooks.spark_sql_hook.SparkSqlHook',
+    ),
 ]
 
 OPERATOR = [
